Penalize gaps within a match more
The scoring of candidates highly favors matches with few gaps. This
commit makes the length of the substring from the first matched
character through the last another factor in determining the cost of a
candidate. Intuitively, this measures the "tightness" of a match. The
effect this has is that some shorter completions can be ordered before
longer ones with fewer gaps. For example, with the search string `ab`:

    a-bx

now gets a lower cost than

    a-xxxxb

even though the latter has one fewer gap.

Implementation-wise, this works by having the cost of a symbol being
in a gap change. When converting Aᵢ to Bⱼ by deleting aᵢ, if j = 0 the
gap will be located before the match, if j = m-1 it will be after, and
otherwise, inside the match.