Sort already in hotfuzz-all-completions
This is a hack since it places additional assumptions on the caller,
although the intended mechanism for sorting is already a hack ugly
enough that it never should have been envisioned in the first place.
By doing things this way, the implementation of the completion style can
be simplified and now calls out to the dynamic module and does
highlighting in a limited capacity.

Closes #2