Autoload hotfuzz--adjust-metadata
An all-completions operation will call the autoloaded function
hotfuzz-all-completions, and as such hotfuzz--adjust-metadata gets
defined before it is subsequently called. However, for try-completion
operations hotfuzz piggybacks on the built-in function
completion-flex-try-completion, and so will not be automatically
loaded. This commit fixes this by also autoloading
hotfuzz--adjust-metadata.

An alternative would have been defining an autoloaded alias for
completion-flex-try-completion, which, though exposing a redundant
symbol, would avoid autoloading a "private" function.